Remove Schema, schemas should inherit from umongo.abstract.BaseSchema
Expose RemoveMissingSchema as umongo.RemoveMissingSchema
Simplify as_marshmallow_schema cache
