+++ tests/test_fields.py
@@ -8,8 +8,9 @@ from bson import ObjectId, DBRef, Decimal128
 import marshmallow as ma
 
 from umongo.data_proxy import data_proxy_factory
-from umongo import Document, EmbeddedDocument, Schema, fields, Reference, validate
+from umongo import Document, EmbeddedDocument, fields, Reference, validate
 from umongo.data_objects import List, Dict
+from umongo.abstract import BaseSchema
 
 from .common import BaseTest
 
@@ -106,7 +107,7 @@ class TestFields(BaseTest):
 
     def test_basefields(self):
 
-        class MySchema(Schema):
+        class MySchema(BaseSchema):
             string = fields.StringField()
             uuid = fields.UUIDField()
             number = fields.NumberField()
@@ -173,7 +174,7 @@ class TestFields(BaseTest):
 
     def test_datetime(self):
 
-        class MySchema(Schema):
+        class MySchema(BaseSchema):
             a = fields.DateTimeField()
 
         s = MySchema()
@@ -205,7 +206,7 @@ class TestFields(BaseTest):
 
         timezone_2h = dt.timezone(dt.timedelta(hours=2), "test")
 
-        class MySchema(Schema):
+        class MySchema(BaseSchema):
             a = fields.AwareDateTimeField()
             b = fields.AwareDateTimeField(default_timezone=timezone_2h)
 
@@ -235,7 +236,7 @@ class TestFields(BaseTest):
 
     def test_date(self):
 
-        class MySchema(Schema):
+        class MySchema(BaseSchema):
             a = fields.DateField()
 
         s = MySchema()
@@ -256,7 +257,7 @@ class TestFields(BaseTest):
 
     def test_dict(self):
 
-        class MySchema(Schema):
+        class MySchema(BaseSchema):
             dict = fields.DictField(attribute='in_mongo_dict', allow_none=True)
             kdict = fields.DictField(keys=fields.StringField(validate=validate.Length(0, 1)))
             vdict = fields.DictField(values=fields.IntField(validate=validate.Range(max=5)))
@@ -334,7 +335,7 @@ class TestFields(BaseTest):
 
     def test_dict_default(self):
 
-        class MySchema(Schema):
+        class MySchema(BaseSchema):
             # Passing a mutable as default is a bad idea in real life
             d_dict = fields.DictField(values=fields.IntField, default={'1': 1, '2': 2})
             c_dict = fields.DictField(values=fields.IntField, default=lambda: {'1': 1, '2': 2})
@@ -422,7 +423,7 @@ class TestFields(BaseTest):
 
     def test_list(self):
 
-        class MySchema(Schema):
+        class MySchema(BaseSchema):
             list = fields.ListField(fields.IntField(), attribute='in_mongo_list', allow_none=True)
 
         MyDataProxy = data_proxy_factory('My', MySchema())
@@ -507,7 +508,7 @@ class TestFields(BaseTest):
 
     def test_list_default(self):
 
-        class MySchema(Schema):
+        class MySchema(BaseSchema):
             d_list = fields.ListField(fields.IntField(), default=(1, 2, 3))
             c_list = fields.ListField(fields.IntField(), default=lambda: (1, 2, 3))
 
@@ -596,7 +597,7 @@ class TestFields(BaseTest):
 
     def test_objectid(self):
 
-        class MySchema(Schema):
+        class MySchema(BaseSchema):
             objid = fields.ObjectIdField(attribute='in_mongo_objid')
 
         MyDataProxy = data_proxy_factory('My', MySchema())
@@ -767,7 +768,7 @@ class TestFields(BaseTest):
 
     def test_decimal(self):
 
-        class MySchema(Schema):
+        class MySchema(BaseSchema):
             price = fields.DecimalField(attribute='in_mongo_price')
 
         MyDataProxy = data_proxy_factory('My', MySchema())
