+++ umongo/expose_missing.py
@@ -3,10 +3,17 @@
 Allows the user to let umongo document return missing rather than None for
 empty fields.
 """
-
-
 from contextvars import ContextVar
 from contextlib import AbstractContextManager
+
+import marshmallow as ma
+
+
+__all__ = (
+    'ExposeMissing',
+    'RemoveMissingSchema',
+)
+
 
 EXPOSE_MISSING = ContextVar("expose_missing", default=False)
 
@@ -24,3 +31,14 @@ class ExposeMissing(AbstractContextManager):
 
     def __exit__(self, *args, **kwargs):
         EXPOSE_MISSING.reset(self.token)
+
+
+class RemoveMissingSchema(ma.Schema):
+    """
+    Custom :class:`marshmallow.Schema` subclass that skips missing fields
+    rather than returning None for missing fields when dumping umongo
+    :class:`umongo.Document`s.
+    """
+    def dump(self, *args, **kwargs):
+        with ExposeMissing():
+            return super().dump(*args, **kwargs)
